Fix unsigned int type errror

+++ pt_mcc/csrc/find_neighbors.cpp
@@ -79,7 +79,7 @@ namespace pt_mcc
 
         unsigned int num_neighbors = std::max(countNeighborsCPU(scale_inv, num_points, num_cells, radius,
                                                                 points_ptr, batch_ids_ptr, points2_ptr, cell_indices_ptr, aabb_min_ptr, aabb_max_ptr, start_indexes_ptr),
-                                              1);
+                                              static_cast<unsigned int>(1));
 
         // Allocate output tensor for neighbor indexes
         torch::Tensor neigh_indexes = torch::zeros({static_cast<int64_t>(num_neighbors), 2}, torch::dtype(torch::kInt32));
